+++ App.xcworkspace/contents.xcworkspacedata
@@ -7,4 +7,7 @@
    <FileRef
       location = "group:">
    </FileRef>
+   <FileRef
+       location = "group:../Fair">
+   </FileRef>
 </Workspace>
