+++ AppFairApp.xcconfig
@@ -7,10 +7,10 @@ PRODUCT_NAME = Net Skip
 PRODUCT_BUNDLE_IDENTIFIER = app.Net-Skip
 
 // The semantic version for the app matching the git tag for the release
-MARKETING_VERSION = 0.0.28
+MARKETING_VERSION = 0.0.29
 
 // The monotonically increasing number specifying the internal app version
-CURRENT_PROJECT_VERSION = 127
+CURRENT_PROJECT_VERSION = 128
 
 // The category of the app should be one of: public.app-category.business, developer-tools, education, entertainment, finance, games, graphics-design, healthcare-fitness, lifestyle, medical, music, news, photography, productivity, reference, social-networking, sports, travel, utilities, video, weather
 APP_CATEGORY = public.app-category.utilities
