+++ Tests/AppTests/AppTests.swift
@@ -14,14 +14,14 @@
  */
 import Swift
 import XCTest
+import FairApp
 @testable import App
 
 open class AppTests: XCTestCase {
-    @available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, *)
-    open func testAppScene() throws {
-        // awaiting Swift 5.5 final
-        //let _ = AppContainer.rootScene
-        //let _ = AppContainer.settingsView
-    }
-}
+    let wv = WebDriver()
 
+    @available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, *)
+    open func testWebState() async throws {
+    }
+
+}
